refactor: testcontainer 설정 infra 모듈로 이동
